Use float function instead.

+++ 579 - Clock Hands.cpp
@@ -11,7 +11,7 @@ int main(void) {
 	cout.setf(ios::fixed, ios::floatfield);
 	while (cin >> h >> c >> m, h | m) {
 		h *= 30;
-		double angle = fabs(h + m * 0.5 - m * 6.0); // (h * 30 + m * 0.5) - (m * 6)
+		float angle = fabs(h + m * 0.5 - m * 6.0); // (h * 30 + m * 0.5) - (m * 6)
 		cout << (angle < 180.0 ? angle : 360.0 - angle) << endl;
 	}
